fix reply when multiple accounts are in replied message
In the original implementation of determine_sender, the first account that
matched a header was taken as the "To:" of the reply. It was broken when the
first account in the list was the account the mail was "Delivered-To:", not
"To:", thus shadowing the account the "To:" header.

This patch fixes the bug using the first header instead of the first account to
get the sender address.